+++ compiler/rustc_mir_build/src/thir/constant.rs
@@ -2,8 +2,7 @@ use rustc_abi::Size;
 use rustc_ast::{self as ast, UintTy};
 use rustc_hir::LangItem;
 use rustc_middle::bug;
-use rustc_middle::mir::interpret::LitToConstInput;
-use rustc_middle::ty::{self, ScalarInt, Ty, TyCtxt, TypeVisitableExt as _};
+use rustc_middle::ty::{self, LitToConstInput, ScalarInt, Ty, TyCtxt, TypeVisitableExt as _};
 use tracing::trace;
 
 use crate::builder::parse_float_into_scalar;
